Move verifier from oak_remote_attestation to  oak_remote_attestation_verification
