+++ oak_remote_attestation_verification/Cargo.toml
@@ -11,8 +11,12 @@ anyhow = "*"
 base64 = "*"
 ecdsa = { version = "*", features = ["pkcs8", "pem"] }
 hex = "*"
+oak_remote_attestation = { workspace = true }
 p256 = { version = "*", features = ["ecdsa-core", "ecdsa", "pem"] }
 serde = { version = "*", features = ["derive"] }
 serde_jcs = "*"
 serde_json = "*"
 sha2 = { version = "*", default-features = false }
+
+[dev-dependencies]
+oak_crypto = { workspace = true }
