+++ oak_remote_attestation_verification/src/lib.rs
@@ -16,7 +16,10 @@
 
 #![cfg_attr(not(test), no_std)]
 
+extern crate alloc;
+
 pub mod rekor;
+pub mod verifier;
 
 #[cfg(test)]
 mod tests;
